add index router

+++ controllers/articles.js
@@ -1,6 +1,8 @@
 const Article = require('../models/article');
 const BadRequestError = require('../errors/BadRequestError');
 const NotFoundError = require('../errors/NotFoundError');
+const ForbiddenError = require('../errors/ForbiddenError');
+
 // GET /articles
 const getArticles = (req, res, next) => {
   Article.find({})
@@ -25,10 +27,13 @@ const createArticle = (req, res, next) => {
 };
 // DELETE /articles/articleId
 const deleteArticle = (req, res, next) => {
-  Article.findById(req.params.articleId)
+  Article.findById(req.params.articleId).select('owner')
     .then((card) => {
       if (!card) {
         throw new NotFoundError('Статья с таким id не найдена.');
+      }
+      if (String(card.owner) !== req.user._id) {
+        throw new ForbiddenError('Нельзя удалять чужие статьи!');
       }
       Article.findByIdAndRemove(req.params.articleId)
         .then(() => res.status(200).send({ message: 'Статья удалена.' }));
